Texture filtering correction is now applied only in the computation the new uv-coordinates.
This corrects programmatic use of the StretchyImage. E.g before that change it was possible to set edgeTop="2" and when reading it via the getter the result was 3!!
This is now applied internally and not part of the members at all.

The texture filtering correction can be toggled internally via a flag (basically const for now). default is true.

------

ResizeableRectangle could also use Modelling.createPlane for creating the base quad. When used properly future widgets could request more complex base shapes (grids) as default. Useful for Stretchy Images and relatives and things like animated meshes (e.g. folding movies via animation nodes)